coord,dataflow: rework log id assignment
The dataflow layer hardcodes several built-in logging sources that are
accessible via SQL. This is somewhat awkward for the coordinator, which
would very much like to be the sole authority on what sources exist in
the system.

The big danger is that both the coord and dataflow crates assign system
IDs, which must never overlap. The current design makes it very hard to
tell what system IDs are in use, and it's easy to accidentally cause
clashes when allocating a new ID.

This commit reworks the dataflow layer to have no opinion on what IDs to
assign to its logging views. The new API looks like so:

    SequencedCommand::EnableLogging(LoggingConfig {
        granularity_ns: u128,
        active_logs: HashMap<LogVariant, GlobalId>,
    })

Essentially, the dataflow layer exposes the log variants it knows how to
maintain, along with an API for installing any subset of the known log
variants with specific IDs. The coordinator is now the sole place where
system IDs are assigned.

A future commit will unify `LogSource`, `LogView`, and `CatalogView`,
which all fall under the umbrella of "built-in system object" and will
be able to share quite a bit of code. The goal here is #913, exposing
`SHOW` data as views, and #2157, supporting pg_catalog.
